refactor: extract numeric helpers
